app/requirements.txt
httpx==0.27.0
boto3==1.34.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-httpx
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc